+++ actionpack/CHANGELOG.md
@@ -4,7 +4,7 @@
 
     *Carlos Galdino + Rafael Mendonça França*
 
-*   Deprecate `:mouseover` options for `image_tag` helper. *Rafael Mendonça França*
+*   Deprecate `:mouseover` option for `image_tag` helper. *Rafael Mendonça França*
 
 *   Deprecate `button_to_function` and `link_to_function` helpers. *Rafael Mendonça França*